zoomcamp/pipeline/assets/ingestion/requirements.txt
# Python dependencies for NYC taxi ingestion (parquet from TLC endpoint).
# Keep minimal; pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python
#
pandas>=2.0.0
pyarrow>=14.0.0
python-dateutil>=2.8.0
tzdata>=2024.1
duckdb>=0.9.0

